idekCTF2025/scanme/attachments/requirements.txt
Flask
PyYAML
python-dotenv
gunicorn
